+++ .gitignore
@@ -34,5 +34,3 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
-
-.fake
