+++ package.json
@@ -9,13 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~4.0.1",
     "expo": "~52.0.33",
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
-    "react-dom": "18.3.1",
     "react-native": "0.76.7",
-    "react-native-web": "~0.19.13"
+    "react-dom": "18.3.1",
+    "react-native-web": "~0.19.13",
+    "@expo/metro-runtime": "~4.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
